[CICD-43] Rename data generator
