+++ lib/data-generators/index.js
@@ -3,5 +3,5 @@ module.exports = {
   "git-tag-commit": require('./git-tag-commit'),
   "git-commit": require('./git-commit'),
   "version-commit": require('./version-commit'),
-  'version-commit-circle': require('./version-commit-circle')
+  'version-commit-circleci': require('./version-commit-circleci')
 };
