+++ lib/data-generators/version-commit-circleci.js
